Keep current directory / dir of file in path
Augmenting 'path' for Vimscript is nice, but I like to keep the default ability to reference files in the current directory and in the current file's directory.
Instead of removing the entire global default value (".,/usr/include,,"), just insert the new paths in place of /usr/include, or append if that path is not there.